+++ packages/snippets/src/toggleThemeViewTransition/index.test.ts
@@ -1,6 +1,6 @@
 import { toggleThemeViewTransition } from '@mixte/snippets/toggleThemeViewTransition';
 import { omit } from 'lodash-es';
-import { deepClone, delay } from 'mixte';
+import { asyncForEach, deepClone, delay } from 'mixte';
 
 describe('toggleThemeViewTransition', () => {
   const mockStartViewTransition = vi.fn(async (fn) => {
@@ -14,8 +14,16 @@ describe('toggleThemeViewTransition', () => {
     };
   });
 
+  let mockMatchMediaMatches = false;
+  const mockMatchMedia = vi.fn(() => {
+    return {
+      matches: mockMatchMediaMatches,
+    };
+  });
+
   let startViewTransition: any;
   let documentElementAnimate: any;
+  let matchMedia: any;
 
   beforeEach(() => { // @ts-expect-error
     startViewTransition = document.startViewTransition;// @ts-expect-error
@@ -25,26 +33,40 @@ describe('toggleThemeViewTransition', () => {
     documentElementAnimate = document.documentElement.animate; // @ts-expect-error
     document.documentElement.animate = mockDocumentElementAnimate;
     mockDocumentElementAnimate.mockClear();
+
+    matchMedia = window.matchMedia; // @ts-expect-error
+    window.matchMedia = mockMatchMedia;
+    mockMatchMediaMatches = false;
+    mockMatchMedia.mockClear();
   });
 
   afterEach(async () => { // @ts-expect-error
     document.startViewTransition = startViewTransition;
     document.documentElement.animate = documentElementAnimate;
+    window.matchMedia = matchMedia;
   });
 
-  it('若浏览器不支持 `document.startViewTransition`, 传入的方法会被直接执行', () => {
+  it('若浏览器不支持 `document.startViewTransition`, 传入的方法会被直接执行', async () => {
     const headChildrenLength = document.head.children.length;
     let isRun = false;
 
     // @ts-expect-error
     document.startViewTransition = undefined;
 
-    toggleThemeViewTransition(() => {
+    const result = toggleThemeViewTransition(() => {
       isRun = true;
     });
 
+    expect(mockStartViewTransition).toHaveBeenCalledTimes(0);
+    expect(mockDocumentElementAnimate).toHaveBeenCalledTimes(0);
     expect(document.head.children.length).toBe(headChildrenLength);
     expect(isRun).toBe(true);
+
+    await result;
+
+    expect(mockStartViewTransition).toHaveBeenCalledTimes(0);
+    expect(mockDocumentElementAnimate).toHaveBeenCalledTimes(0);
+    expect(document.head.children.length).toBe(headChildrenLength);
   });
 
   it('若浏览器支持 `document.startViewTransition`, 执行正常流程', async () => {
@@ -62,7 +84,9 @@ describe('toggleThemeViewTransition', () => {
 
     await nextTick();
 
+    expect(mockStartViewTransition).toHaveBeenCalledTimes(1);
     expect(mockDocumentElementAnimate).toHaveBeenCalledTimes(1);
+    expect(document.head.children.length).toBe(headChildrenLength + 1);
 
     await result;
 
@@ -129,5 +153,48 @@ describe('toggleThemeViewTransition', () => {
     expect(noClipPath[1].pseudoElement).toBe('::view-transition-new(root)');
     expect(trueClipPath[1].pseudoElement).toBe('::view-transition-old(root)');
     expect(falseClipPath[1].pseudoElement).toBe('::view-transition-new(root)');
+  });
+
+  it('传入 prefersReducedMotion 选项时, 若设置了偏好减少动画则不进行动画, 传入的方法会被直接执行', async () => {
+    const headChildrenLength = document.head.children.length;
+    let isRun = false;
+
+    await asyncForEach([
+      [undefined, false] as const,
+      [true, false] as const,
+      [false, false] as const,
+      [undefined, true] as const,
+      [true, true] as const,
+      [false, true] as const,
+    ], async ([prefersReducedMotion, _mockMatchMediaMatches]) => {
+      isRun = false;
+      mockMatchMediaMatches = _mockMatchMediaMatches;
+      mockMatchMedia.mockClear();
+
+      const result = toggleThemeViewTransition(() => {
+        isRun = true;
+      }, {
+        prefersReducedMotion,
+      });
+
+      // 设置了检测 ( 不传时默认检测 ), 那么就会执行 window.matchMedia 方法
+      expect(mockMatchMedia).toBeCalledTimes([undefined, true].includes(prefersReducedMotion) ? 1 : 0);
+      // 测试是否执行了动画, 通过是否插入了样式来判断
+      expect(document.head.children.length).toBe(
+        // 设置了检测
+        [undefined, true].includes(prefersReducedMotion)
+          // 检测通过, 那就会执行动画
+          ? !mockMatchMediaMatches ? headChildrenLength + 1 : headChildrenLength
+          // 设置了不检测, 无论检测是否通过, 都会执行动画
+          : headChildrenLength + 1,
+      );
+      // 无论如何, 传入的方法都会被执行
+      expect(isRun).toBe(true);
+
+      await result;
+
+      // 插入的样式都会被重置
+      expect(document.head.children.length).toBe(headChildrenLength);
+    });
   });
 });